+++ quna/src/main/java/com/ktpt/quna/infra/QuestionQueryDAO.java
@@ -1,4 +1,4 @@
-package com.ktpt.quna.domain.model;
+package com.ktpt.quna.infra;
 
 import static com.ktpt.quna.domain.model.QMember.*;
 import static com.ktpt.quna.domain.model.QQuestion.*;
@@ -7,14 +7,15 @@ import java.util.List;
 
 import org.springframework.stereotype.Repository;
 
+import com.ktpt.quna.domain.model.Question;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 @Repository
-public class QuestionQueryDao {
+public class QuestionQueryDAO {
 
 	private final JPAQueryFactory query;
 
-	public QuestionQueryDao(JPAQueryFactory query) {
+	public QuestionQueryDAO(JPAQueryFactory query) {
 		this.query = query;
 	}
